# lib/api-zod/src/generated/types/labEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Bloodchain Platform API — Strand ledger + Torrent routing engine
 * OpenAPI spec version: 1.0.0
 */

export interface LabEvent {
  eventId: string;
  unitId: string;
  facilityId: string;
  action: string;
  actor: string;
  /** @nullable */
  reason?: string | null;
  timestamp: Date;
  chainHash: string;
}