# lib/api-zod/src/generated/types/registerDonorInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Bloodchain Platform API — Strand ledger + Torrent routing engine
 * OpenAPI spec version: 1.0.0
 */

export interface RegisterDonorInput {
  name: string;
  phone: string;
  bloodType: string;
  district: string;
  email?: string;
  idNumber?: string;
}